+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Additional labels for pre-release and build metadata are available as extensions
 10. Updated functionality in `plot_timeseries` function (Issue [#457](https://github.com/brightwind-dev/brightwind/issues/457)).
 11. More representative error messages returned in `Shear` function when no valid data is available (Issue [#205](https://github.com/brightwind-dev/brightwind/issues/205)).
 11. `Shear.TimeSeries` and `Shear.TimeOfDay` has been updated to return np.nan when there is no valid data for that timestamp or hour (Issue [#205](https://github.com/brightwind-dev/brightwind/issues/205)).
+12. Updated functionality in `get_measurement_stations` to enable filtering for device type (Issue [#378](https://github.com/brightwind-dev/brightwind/issues/378)).
 12. Added new function `LoadBrightHub.get_cleaning_rules()` to pull the cleaning rules for a particular measurement station on BrightHub (Issue [#461](https://github.com/brightwind-dev/brightwind/issues/461)).
 13. Added new function `load.apply_cleaning_rules()` to clean columns specified in cleaning rules by the condition specified for a specific column in the rules. (Issue [#462](https://github.com/brightwind-dev/brightwind/issues/462)).
 14. Fix bug for application of `Shear.TimeOfDay()` (Issue [#441](https://github.com/brightwind-dev/brightwind/issues/441)).